requirements.txt
altair==5.1.2
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Django==4.2.6
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-livereload-server==0.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
gitdb==4.0.10
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mssql-django==1.3
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
plotly==5.17.0
protobuf==4.24.4
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
pyodbc==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.4.4
streamlit==1.27.2
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
